Move TimestampDomain into its own file / submodule
